Custom Icon toasts (#475)
Extend CustomToast with icon
