+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- [Added icon functionality to custom toast](https://github.com/ElrondNetwork/dapp-core/pull/475)
 ## [[2.1.11](https://github.com/ElrondNetwork/dapp-core/pull/474)] - 2022-11-15
 - [Changed `ConfirmationScreen` calling sign hooks depeding on device](https://github.com/ElrondNetwork/dapp-core/pull/461)
 - [Updated legacy peers type definitions](https://github.com/ElrondNetwork/dapp-core/pull/460)
