+++ src/UI/TransactionDetails/transactionDetails.styles.scss
