+++ src/UI/TransactionsToastList/components/CustomToast/components/index.tsx
@@ -1,0 +1,2 @@
+export * from './SimpleToast';
+export * from './IconToast';
