+++ src/UI/TransactionsToastList/components/CustomToast/customToast.styles.scss
