+++ src/UI/TransactionsToastList/components/TransactionToast/components/DefaultToastDeleteButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { faTimes } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { WithClassnameType } from '../../../types';
+import { WithClassnameType } from '../../../../types';
 
 export interface DefaultToastDeleteButtonPropsType extends WithClassnameType {
   onClick?: () => void;
