+++ src/UI/TransactionsToastList/components/TransactionToast/components/index.tsx
@@ -1,0 +1,3 @@
+export * from './DefaultToastDeleteButton';
+export * from './TransactionToastContent';
+export * from './TransactionToastWrapper';
