+++ src/UI/TransactionsToastList/components/TransactionToast/hooks/useTransactionToast.ts
@@ -1,24 +1,21 @@
 import { useEffect, useMemo } from 'react';
+import { AVERAGE_TX_DURATION_MS, CROSS_SHARD_ROUNDS } from 'constants/index';
 import { useGetTransactionDisplayInfo } from 'hooks';
 import { useSelector } from 'reduxStore/DappProviderContext';
 import { shardSelector } from 'reduxStore/selectors';
 import { TransactionBatchStatusesEnum } from 'types';
 
+import { getUnixTimestamp } from 'utils/dateTime/getUnixTimestamp';
+import { getUnixTimestampWithAddedMilliseconds } from 'utils/dateTime/getUnixTimestampWithAddedMilliseconds';
 import { getAreTransactionsOnSameShard } from 'utils/transactions/getAreTransactionsOnSameShard';
 import {
   getIsTransactionPending,
   getIsTransactionTimedOut
 } from 'utils/transactions/transactionStateByStatus';
 
-import { getUnixTimestamp } from 'utils/dateTime/getUnixTimestamp';
-import { getUnixTimestampWithAddedMilliseconds } from 'utils/dateTime/getUnixTimestampWithAddedMilliseconds';
-
-import styles from '../styles.scss';
+import styles from '../transactionToast.styles.scss';
 import { TransactionToastDefaultProps } from '../transactionToast.type';
 import { getToastDataStateByStatus } from '../utils';
-
-const AVERAGE_TX_DURATION_MS = 6000;
-const CROSS_SHARD_ROUNDS = 5;
 
 export const useTransactionToast = ({
   toastId,
