+++ src/UI/TransactionsToastList/components/TransactionToast/transactionToast.styles.scss
