--- src/UI/TransactionsToastList/components/index.ts
@@ -1,2 +1,0 @@
-export { TransactionToast } from './TransactionToast/index';
-export { CustomToast } from './CustomToast';
