+++ src/UI/TransactionsToastList/transactionsToastList.styles.scss
