+++ src/constants/transactionStatus.ts
@@ -1,1 +1,3 @@
 export const TRANSACTION_STATUS_TOAST_ID = 'transaction-status-toast-id';
+export const AVERAGE_TX_DURATION_MS = 6000;
+export const CROSS_SHARD_ROUNDS = 5;
