+++ src/utils/toasts/customToastsActions.ts
@@ -1,9 +1,9 @@
 import { addCustomToast, removeCustomToast } from 'reduxStore/slices';
 import { store } from 'reduxStore/store';
-import { CustomToastType } from 'types/toasts.types';
+import { CustomToastType } from '../../types/toasts.types';
 
-export const addNewCustomToast = (args: CustomToastType) => {
-  return store.dispatch(addCustomToast(args)).payload;
+export const addNewCustomToast = (props: CustomToastType) => {
+  return store.dispatch(addCustomToast(props)).payload;
 };
 
 export const deleteCustomToast = (toastId: string) =>
